+++ src/utils/low_pass_filter.h
@@ -9,7 +9,7 @@ class LowPassFilter {
    public:
     LowPassFilter(float cut_off_frequency, float delta_time)
         : value_{}, e_pow_{0.0F}, first_value_{true} {
-        e_pow_ = 1 - std::exp(-delta_time * 2 * M_PI * cut_off_frequency);
+        e_pow_ = 1 - expf(-delta_time * 2.0f * M_PI * cut_off_frequency);
     }
     T update(T input) {
         if (first_value_) {
